Use vector<int> for fload_indexed

+++ lib/include/resdata/FortIO.hpp
@@ -77,7 +77,7 @@ public:
     bool fseek(offset_type offset, int whence);
     bool data_fskip(int element_size, int element_count, int block_count);
     void data_fseek(offset_type data_offset, size_t data_element,
-                    int element_size, int element_count, int block_size);
+                    size_t element_size, int element_count, int block_size);
     bool ftruncate(offset_type size);
     int fclean();
     bool fclose_stream();
